Fix boot on Raspberry pi
With the changes we made to use fit images, `kernel_image` refers to the
full FIT image, to be unpacked by u-boot. It's large and u-boot will
extract the kernel from there to the correct location, so there is no
need to load it at kernel_addr_r.

Loading it there made some blobs overlap and the device could not boot.