Revert "u-boot: refresh patches for 2019.10."
This reverts commit 18da5914a091c97f3a5c1c53f5be8eee45eed1a5.

This should not have been backported to zeus; poky in zeus is still
using u-boot 2019.07. This fixes some bitbake warnings about fuzz
applying patches. Note that there are still some warnings remaining, but
they are because of patches in meta-raspberrypi. Those are pretty
trivial so I'm not worried about that.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>